refactor SelectProjectsDialog controller concerns * clicking "Continue to scope selection" moves to next section AND automatically opens the project selection dialog
