+++ app/controllers/admin/import/jira/import_runs_controller.rb
@@ -32,6 +32,7 @@ module Admin::Import::Jira
   class ImportRunsController < ApplicationController
     include OpTurbo::ComponentStream
     include ImportRuns::ComponentStreams
+    include ImportRuns::SelectProjectsDialog
 
     layout "admin"
 
@@ -58,8 +59,11 @@ module Admin::Import::Jira
     end
 
     def continue
-      change_step(params[:step]) unless @jira_import.status_running?
-      stream_wizard
+      step = params[:step]
+      change_step(step) unless @jira_import.status_running?
+      stream_wizard do
+        open_select_projects_dialog if step.present? && step.to_sym == :configure && @jira_import.in_state?(:configuring)
+      end
     rescue StandardError => e
       handle_error(e)
     end
